// src/app/model/employee.model.ts
export interface Employees {
  id: number;
  firstName: string;
  lastName: string;
  gender: string;
  dateOfBirth: string;
  dateOfJoining: string;
  phoneNumber: string;
  emailId: string;
  adderssLine1: string;
  adderssLine2: string;
  street: string;
  district: string;
  state: string;
  country: string;
  pinCode: string;
}